Confidence Intervals for pseudoglm Model Parameters

Computes Wald confidence intervals for one or more parameters in a fitted model. Users can specify the type of variance estimate used, with the default being the robust sandwich variance estimator.

Usage

# S3 method for class 'pseudoglm'
confint(object, parm, level = 0.95, type = "robust", ...)

Arguments

object

A fitted model object from cumincglm or rmeanglm

parm

a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

the confidence level required.

type

The type of variance estimate to use, see vcov.pseudoglm

...

Not used

Value

A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in

Examples

cumincipcw <- cumincglm(survival::Surv(etime, event) ~ age + sex,
         time = 200, cause = "pcm", link = "identity",
         model.censoring = "independent", data = mgus2)
confint(cumincipcw)
#>                    2.5 %        97.5 %
#> (Intercept)  0.118603220  0.3264884927
#> age         -0.003125601 -0.0004437915
#> sexM        -0.043348939  0.0251469048