Compute covariance matrix of regression coefficient estimates

Compute covariance matrix of regression coefficient estimates

# S3 method for pseudoglm
vcov(object, type = "robust", ...)

Arguments

object

A pseudoglm object, as returned by cumincglm or rmeanglm.

type

The method to use for variance estimation; one of "corrected", "robust", "naive", or "cluster"

...

Not used

Value

A numeric matrix containing the variance-covariance estimates

Details

The "corrected" variance estimate for the cumulative incidence is as described in Overgaard et al. (2017) <doi:10.1214/16-AOS1516>, with code adapted from Overgaard's Stata program. This method does not handle ties and only has marginal benefits in reasonable sample sizes. The default is "robust" which uses the sandwich estimator vcovHC as implemented in the sandwich package. "cluster" is another option if you have clustered observations that uses the vcovCL function in sandwich. Finally "naive" uses the same method as glm to compute the variance, and is known to be anti-conservative. The bootstrap is another recommended option that can be implemented using other tools; there is an example in the vignette.

References

Overgaard, Morten; Parner, Erik Thorlund; Pedersen, Jan. Asymptotic theory of generalized estimating equations based on jack-knife pseudo-observations. Ann. Statist. 45 (2017), no. 5, 1988--2015. <doi:10.1214/16-AOS1516>.

See also