Compute pseudo observations under independent censoring

Assuming completely independent censoring, i.e., censoring does not depend on the survival time nor any covariates in the model, the pseudo observations are calculated with the standard jackknife approach

Usage

pseudo_independent(
  formula,
  time,
  cause = 1,
  data,
  type = c("cuminc", "survival", "rmean"),
  formula.censoring = NULL,
  ipcw.method = NULL
)

Arguments

formula

A formula specifying the model. The left hand side must be a Surv object specifying a right censored survival or competing risks outcome. The status indicator, normally 0=alive, 1=dead. Other choices are TRUE/FALSE (TRUE = death) or 1/2 (2=death). For competing risks, the event variable will be a factor, whose first level is treated as censoring. The right hand side is the usual linear combination of covariates.

time

Numeric constant specifying the time at which the cumulative incidence or survival probability effect estimates are desired.

cause

Numeric or character constant specifying the cause indicator of interest.

data

Data frame in which all variables of formula can be interpreted.

type

One of "survival", "cuminc", or "rmean"

formula.censoring

Not used with this method, see pseudo_stratified, pseudo_aareg or pseudo_coxph

ipcw.method

Not used with this method

Value

A vector of jackknife pseudo observations

Examples

POi <- pseudo_independent(Surv(time, status) ~ 1, 1500, cause = 1, data = colon, type = "survival")
mean(POi)
#> [1] 0.5980774